+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 
 # Changelog
 
-#### 29th October 2024 - Version 2.0.16
+#### 29th October 2024 - Version 2.0.17
 
 -   Quickfix: Reenqueued album disc JS script with wp_enqueue_script function to fix missing animation
 -   Improvement: Added error log to be able to debug issue with activation
+-   Fix: Default Google Maps height
 
 #### 28th October 2024 - Version 2.0.14